# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: الدمشقي ERP + e-commerce + POS + manufacturing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pos/daily-closing/{date}:
    get:
      operationId: getDailyClosing
      tags: [pos]
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyClosing" }
    post:
      operationId: closeDailyClosing
      tags: [pos]
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                countedCash: { type: integer }
                note: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DailyClosing" }
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Employee" }
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /employees/{id}/attendance:
    get:
      operationId: listAttendance
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: month
          in: query
          required: false
          schema: { type: string, description: "YYYY-MM" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AttendanceRecord" }
    post:
      operationId: createAttendance
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendanceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttendanceRecord" }
  /employees/salaries:
    get:
      operationId: listSalaries
      tags: [employees]
      parameters:
        - name: month
          in: query
          required: false
          schema: { type: string, description: "YYYY-MM" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SalaryRecord" }
    post:
      operationId: paySalary
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId, month, amount]
              properties:
                employeeId: { type: string }
                month: { type: string }
                amount: { type: integer }
                bonus: { type: integer }
                deduction: { type: integer }
                note: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalaryRecord" }
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MediaAsset" }
    post:
      operationId: createMedia
      tags: [media]
      summary: Register an uploaded object as a media asset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MediaInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaAsset" }
  /media/{id}:
    delete:
      operationId: deleteMedia
      tags: [media]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    DailyClosing:
      type: object
      required: [date, completedCount, salesTotalMinor, cashTotalMinor, cardTotalMinor, isClosed]
      properties:
        date: { type: string, format: date }
        completedCount: { type: integer }
        salesTotalMinor: { type: integer }
        cashTotalMinor: { type: integer }
        cardTotalMinor: { type: integer }
        otherTotalMinor: { type: integer }
        breakdown:
          type: object
          additionalProperties: { type: integer }
          description: Sales totals by payment method (cash, card, cod, bank_transfer, stripe, paypal)
        countedCashMinor: { type: integer }
        varianceMinor: { type: integer }
        isClosed: { type: boolean }
        closedAt: { type: string, format: date-time }
        note: { type: string }
...
    Employee:
      type: object
      required: [id, fullNameAr, role, baseSalaryMinor, currency, hiredOn, isActive, createdAt]
      properties:
        id: { type: string }
        fullNameAr: { type: string }
        role: { type: string }
        phone: { type: string }
        nationalId: { type: string }
        baseSalaryMinor: { type: integer }
        currency: { type: string }
        hiredOn: { type: string, format: date }
        notesAr: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    EmployeeInput:
      type: object
      required: [fullNameAr, role, baseSalaryMinor, hiredOn]
      properties:
        fullNameAr: { type: string }
        role: { type: string }
        phone: { type: string }
        nationalId: { type: string }
        baseSalaryMinor: { type: integer }
        hiredOn: { type: string, format: date }
        notesAr: { type: string }
        isActive: { type: boolean }
    AttendanceRecord:
      type: object
      required: [id, employeeId, date, status, createdAt]
      properties:
        id: { type: string }
        employeeId: { type: string }
        date: { type: string, format: date }
        status: { type: string, enum: [present, absent, late, leave] }
        hoursWorked: { type: integer }
        note: { type: string }
        createdAt: { type: string, format: date-time }
    AttendanceInput:
      type: object
      required: [date, status]
      properties:
        date: { type: string, format: date }
        status: { type: string, enum: [present, absent, late, leave] }
        hoursWorked: { type: integer }
        note: { type: string }
    SalaryRecord:
      type: object
      required: [id, employeeId, employeeNameAr, month, baseAmountMinor, totalAmountMinor, currency, isPaid, createdAt]
      properties:
        id: { type: string }
        employeeId: { type: string }
        employeeNameAr: { type: string }
        month: { type: string }
        baseAmountMinor: { type: integer }
        bonusMinor: { type: integer }
        deductionMinor: { type: integer }
        totalAmountMinor: { type: integer }
        currency: { type: string }
        isPaid: { type: boolean }
        paidAt: { type: string, format: date-time }
        note: { type: string }
        createdAt: { type: string, format: date-time }
...
    MediaAsset:
      type: object
      required: [id, name, url, kind, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        url: { type: string }
        objectPath: { type: string }
        kind: { type: string, enum: [image, document] }
        sizeBytes: { type: integer }
        createdAt: { type: string, format: date-time }
    MediaInput:
      type: object
      required: [name, uploadURL]
      properties:
        name: { type: string }
        uploadURL: { type: string }
        kind: { type: string, enum: [image, document] }
...